+++ addons/units/CfgVehicles/base_item.hpp
@@ -1,0 +1,8 @@
+#define MACRO_ITEM_COMMON \
+    dlc = QUOTE(PREFIX); \
+    scope = 2; \
+    scopeCurator = 2; \
+    vehicleClass = "ItemsUniforms"; \
+    model = "\A3\Weapons_F\DummySuitpack.p3d";
+
+class Item_Base_F;
