+++ addons/units/CfgVehicles/combat_long_helpers.hpp
@@ -1,0 +1,86 @@
+// Naming pattern: SizeSleeves_ColorShirt_ColorPants_ColorBoots
+// Combat - Long Sleeves
+class CLASS(Unit_Combat_LS_BS_BP_BB): CLASS(Unit_Combat_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Kresky";
+    uniformClass = QCLASS(Uniform_Combat_LS_BS_BP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_bs_bp_bb_co.paa)};
+};
+class CLASS(Unit_Combat_LS_BS_GP_BB): CLASS(Unit_Combat_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Combat_LS_BS_GP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_bs_gp_bb_co.paa)};
+};
+class CLASS(Unit_Combat_LS_BS_GP_TB): CLASS(Unit_Combat_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Combat_LS_BS_GP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_bs_gp_tb_co.paa)};
+};
+class CLASS(Unit_Combat_LS_BS_TP_BB): CLASS(Unit_Combat_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Combat_LS_BS_TP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_bs_tp_bb_co.paa)};
+};
+class CLASS(Unit_Combat_LS_BS_TP_TB): CLASS(Unit_Combat_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Combat_LS_BS_TP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_bs_tp_tb_co.paa)};
+};
+class CLASS(Unit_Combat_LS_GS_GP_BB): CLASS(Unit_Combat_LS_Base) {
+    scope = 1;
+    author = "Kresky";
+    uniformClass = QCLASS(Uniform_Combat_LS_GS_GP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_gs_gp_bb_co.paa)};
+};
+class CLASS(Unit_Combat_LS_GS_BP_BB): CLASS(Unit_Combat_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Combat_LS_GS_BP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_gs_bp_bb_co.paa)};
+};
+class CLASS(Unit_Combat_LS_GS_TP_BB): CLASS(Unit_Combat_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Combat_LS_GS_TP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_gs_tp_bb_co.paa)};
+};
+class CLASS(Unit_Combat_LS_BS_DGP_BB): CLASS(Unit_Combat_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Combat_LS_BS_DGP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_bs_dgp_bb_co.paa)};
+};
+class CLASS(Unit_Combat_LS_TS_BP_BB): CLASS(Unit_Combat_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Combat_LS_TS_BP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_ts_bp_bb_co.paa)};
+};
+class CLASS(Unit_Combat_LS_TS_GP_BB): CLASS(Unit_Combat_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Combat_LS_TS_GP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_ts_gp_bb_co.paa)};
+};
+class CLASS(Unit_Combat_LS_TS_GP_TB): CLASS(Unit_Combat_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Combat_LS_TS_GP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_ts_gp_tb_co.paa)};
+};
+class CLASS(Unit_Combat_LS_TS_TP_TB): CLASS(Unit_Combat_LS_Base) {
+    scope = 1;
+    author = "Kresky";
+    uniformClass = QCLASS(Uniform_Combat_LS_TS_TP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_ts_tp_tb_co.paa)};
+};
+class CLASS(Unit_Combat_LS_WS_WP_WB): CLASS(Unit_Combat_LS_Base) {
+    scope = 1;
+    author = "Kresky";
+    uniformClass = QCLASS(Uniform_Combat_LS_WS_WP_WB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_ws_wp_wb_co.paa)};
+};
