+++ addons/units/CfgVehicles/combat_long_item.hpp
@@ -1,0 +1,126 @@
+class CLASS(Item_Uniform_Combat_LS_BS_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Kresky";
+    displayName = CSTRING(Uniform_Combat_LS_BS_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_BS_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_BS_GP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_BS_GP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_BS_GP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_BS_GP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_BS_GP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_BS_GP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_BS_TP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_BS_TP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_BS_TP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_BS_TP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_BS_TP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_BS_TP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_GS_GP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_GS_GP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_GS_GP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_GS_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_GS_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_GS_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_GS_TP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_GS_TP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_GS_TP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_BS_DGP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_BS_DGP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_BS_DGP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_TS_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_TS_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_TS_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_TS_GP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_TS_GP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_TS_GP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_TS_GP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_TS_GP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_TS_GP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_TS_TP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_TS_TP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_TS_TP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_WS_WP_WB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas, Kresky";
+    displayName = CSTRING(Uniform_Combat_LS_WS_WP_WB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_WS_WP_WB),1);
+    };
+};
