+++ addons/units/CfgVehicles/combat_long_plaid_item.hpp
@@ -1,0 +1,81 @@
+class CLASS(Item_Uniform_Combat_LS_CBPS_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "GilleeDoo, Tyrone";
+    displayName = CSTRING(Uniform_Combat_LS_CBPS_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CBPS_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_CDES_TP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "GilleeDoo, Tyrone";
+    displayName = CSTRING(Uniform_Combat_LS_CDES_TP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CDES_TP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_CDFS_GP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "GilleeDoo, Tyrone";
+    displayName = CSTRING(Uniform_Combat_LS_CDFS_GP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CDFS_GP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_CLES_TP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "GilleeDoo, Tyrone";
+    displayName = CSTRING(Uniform_Combat_LS_CLES_TP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CLES_TP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_CLFS_GP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "GilleeDoo, Tyrone";
+    displayName = CSTRING(Uniform_Combat_LS_CLFS_GP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CLFS_GP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_CDBS_GP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_CDBS_GP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CDBS_GP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_CLBS_GP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_CLBS_GP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CLBS_GP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_CLRS_TP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_CLRS_TP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CLRS_TP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_LS_CPS_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_CPS_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CPS_BP_BB),1);
+    };
+};
