+++ addons/units/CfgVehicles/combat_rolled_atacs_helpers.hpp
@@ -1,0 +1,26 @@
+// Naming pattern: SleeveType_Camo_Pattern
+// Combat ATACS - Short Sleeves
+class CLASS(Unit_Combat_RS_ATACS_ATX): CLASS(Unit_Combat_RS_Base) {
+    scope = 1;
+    author = "Tibernius, Tyrone";
+    uniformClass = QCLASS(Uniform_Combat_RS_ATACS_ATX);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_atacs_atx_co.paa)};
+};
+class CLASS(Unit_Combat_RS_ATACS_AUX): CLASS(Unit_Combat_RS_Base) {
+    scope = 1;
+    author = "Tibernius, Tyrone";
+    uniformClass = QCLASS(Uniform_Combat_RS_ATACS_AUX);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_atacs_aux_co.paa)};
+};
+class CLASS(Unit_Combat_RS_ATACS_IX): CLASS(Unit_Combat_RS_Base) {
+    scope = 1;
+    author = "Tibernius, Tyrone";
+    uniformClass = QCLASS(Uniform_Combat_RS_ATACS_IX);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_atacs_ix_co.paa)};
+};
+class CLASS(Unit_Combat_RS_ATACS_LEX): CLASS(Unit_Combat_RS_Base) {
+    scope = 1;
+    author = "Tibernius, Tyrone";
+    uniformClass = QCLASS(Uniform_Combat_RS_ATACS_LEX);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_atacs_lex_co.paa)};
+};
