+++ addons/units/CfgVehicles/combat_rolled_atacs_item.hpp
@@ -1,0 +1,36 @@
+class CLASS(Item_Uniform_Combat_RS_ATACS_ATX): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Tibernius, Tyrone";
+    displayName = CSTRING(Uniform_Combat_RS_ATACS_ATX);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_ATACS_ATX),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_RS_ATACS_AUX): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Tibernius, Tyrone";
+    displayName = CSTRING(Uniform_Combat_RS_ATACS_AUX);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_ATACS_AUX),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_RS_ATACS_IX): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Tibernius, Tyrone";
+    displayName = CSTRING(Uniform_Combat_RS_ATACS_IX);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_ATACS_IX),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_RS_ATACS_LEX): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Tibernius, Tyrone";
+    displayName = CSTRING(Uniform_Combat_RS_ATACS_LEX);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_ATACS_LEX),1);
+    };
+};
