+++ addons/units/CfgVehicles/combat_rolled_item.hpp
@@ -1,0 +1,117 @@
+class CLASS(Item_Uniform_Combat_RS_BS_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas, Kresky";
+    displayName = CSTRING(Uniform_Combat_RS_BS_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_BS_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_RS_BS_GP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_BS_GP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_BS_GP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_RS_BS_GP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_BS_GP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_BS_GP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_RS_BS_TP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_BS_TP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_BS_TP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_RS_BS_TP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_BS_TP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_BS_TP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_RS_GS_GP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_GS_GP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_GS_GP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_RS_GS_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_GS_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_GS_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_RS_GS_TP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_GS_TP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_GS_TP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_RS_BS_DGP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_BS_DGP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_BS_DGP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_RS_TS_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_TS_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_TS_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_RS_TS_GP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_TS_GP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_TS_GP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_RS_TS_GP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_TS_GP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_TS_GP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Combat_RS_TS_TP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_TS_TP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_TS_TP_TB),1);
+    };
+};
