+++ addons/units/CfgVehicles/combat_rolled_plaid_helpers.hpp
@@ -1,0 +1,56 @@
+// Naming pattern: SizeSleeves_ColorShirt_ColorPants_ColorBoots
+// Combat - Rolled Sleeves - Plaid
+class CLASS(Unit_Combat_RS_CBPS_BP_BB): CLASS(Unit_Combat_RS_C_Base) {
+    scope = 1;
+    author = "GilleeDoo, Tyrone";
+    uniformClass = QCLASS(Uniform_Combat_RS_CBPS_BP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_cbps_bp_bb_co.paa)};
+};
+class CLASS(Unit_Combat_RS_CDES_TP_TB): CLASS(Unit_Combat_RS_C_Base) {
+    scope = 1;
+    author = "GilleeDoo, Tyrone";
+    uniformClass = QCLASS(Uniform_Combat_RS_CDES_TP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_cdes_tp_tb_co.paa)};
+};
+class CLASS(Unit_Combat_RS_CDFS_GP_TB): CLASS(Unit_Combat_RS_C_Base) {
+    scope = 1;
+    author = "GilleeDoo, Tyrone";
+    uniformClass = QCLASS(Uniform_Combat_RS_CDFS_GP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_cdfs_gp_tb_co.paa)};
+};
+class CLASS(Unit_Combat_RS_CLES_TP_BB): CLASS(Unit_Combat_RS_C_Base) {
+    scope = 1;
+    author = "GilleeDoo, Tyrone";
+    uniformClass = QCLASS(Uniform_Combat_RS_CLES_TP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_cles_tp_bb_co.paa)};
+};
+class CLASS(Unit_Combat_RS_CLFS_GP_BB): CLASS(Unit_Combat_RS_C_Base) {
+    scope = 1;
+    author = "GilleeDoo, Tyrone";
+    uniformClass = QCLASS(Uniform_Combat_RS_CLFS_GP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_clfs_gp_bb_co.paa)};
+};
+class CLASS(Unit_Combat_RS_CDBS_GP_TB): CLASS(Unit_Combat_RS_C_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Combat_RS_CDBS_GP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_cdbs_gp_tb_co.paa)};
+};
+class CLASS(Unit_Combat_RS_CLBS_GP_BB): CLASS(Unit_Combat_RS_C_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Combat_RS_CLBS_GP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_clbs_gp_bb_co.paa)};
+};
+class CLASS(Unit_Combat_RS_CLRS_TP_BB): CLASS(Unit_Combat_RS_C_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Combat_RS_CLRS_TP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_clrs_tp_bb_co.paa)};
+};
+class CLASS(Unit_Combat_RS_CPS_BP_BB): CLASS(Unit_Combat_RS_C_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Combat_RS_CPS_BP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_combat_cps_bp_bb_co.paa)};
+};
