+++ addons/units/CfgVehicles/floral_jeans_helpers.hpp
@@ -1,0 +1,42 @@
+// Naming pattern: SizeSleeves_ColorShirt_ColorPants_ColorBoots
+// Floral Shirts with Jeans
+class CLASS(Unit_Floral_JP_BS_LP_BB): CLASS(Unit_TShirt_Floral_Denim_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Floral_JP_BS_LP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_floral_denim_bs_bp_bb.paa)
+    };
+};
+class CLASS(Unit_Floral_JP_RS_LP_BB): CLASS(Unit_TShirt_Floral_Denim_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Floral_JP_RS_LP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_floral_denim_rs_bp_bb.paa)
+    };
+};
+class CLASS(Unit_Floral_JP_GS_LP_BB): CLASS(Unit_TShirt_Floral_Denim_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Floral_JP_GS_LP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_floral_denim_gs_bp_bb.paa)
+    };
+};
+class CLASS(Unit_Floral_JP_WS_LP_BB): CLASS(Unit_TShirt_Floral_Denim_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Floral_JP_GS_LP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_floral_denim_ws_bp_bb.paa)
+    };
+};
+class CLASS(Unit_Floral_JP_BYS_LP_BB): CLASS(Unit_TShirt_Floral_Denim_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Floral_JP_BYS_LP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_floral_denim_bys_bp_bb.paa)
+    };
+};
