+++ addons/units/CfgVehicles/floral_jeans_item.hpp
@@ -1,0 +1,45 @@
+class CLASS(Item_Uniform_Floral_JP_BS_LP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Floral_JP_BS_LP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Floral_JP_BS_LP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Floral_JP_RS_LP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Floral_JP_RS_LP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Floral_JP_RS_LP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Floral_JP_GS_LP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Floral_JP_GS_LP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Floral_JP_GS_LP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Floral_JP_WS_LP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Floral_JP_WS_LP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Floral_JP_WS_LP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Floral_JP_BYS_LP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Floral_JP_BYS_LP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Floral_JP_BYS_LP_BB),1);
+    };
+};
