+++ addons/units/CfgVehicles/floral_shorts_helpers.hpp
@@ -1,0 +1,42 @@
+// Naming pattern: SizeSleeves_ColorShirt_ColorPants_ColorBoots
+// Floral Shirts with Shorts
+class CLASS(Unit_Floral_SP_BS_LP_BB): CLASS(Unit_TShirt_Floral_Shorts_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Floral_SP_BS_LP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_floral_shorts_bs_bp_bb.paa)
+    };
+};
+class CLASS(Unit_Floral_SP_RS_LP_BB): CLASS(Unit_TShirt_Floral_Shorts_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Floral_SP_RS_LP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_floral_shorts_rs_bp_bb.paa)
+    };
+};
+class CLASS(Unit_Floral_SP_GS_LP_BB): CLASS(Unit_TShirt_Floral_Shorts_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Floral_SP_GS_LP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_floral_shorts_gs_bp_bb.paa)
+    };
+};
+class CLASS(Unit_Floral_SP_WS_LP_BB): CLASS(Unit_TShirt_Floral_Shorts_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Floral_SP_WS_LP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_floral_shorts_ws_bp_bb.paa)
+    };
+};
+class CLASS(Unit_Floral_SP_BYS_LP_BB): CLASS(Unit_TShirt_Floral_Shorts_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Floral_SP_BYS_LP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_floral_shorts_bys_bp_bb.paa)
+    };
+};
