+++ addons/units/CfgVehicles/garment_long_helpers.hpp
@@ -1,0 +1,86 @@
+// Naming pattern: SizeSleeves_ColorShirt_ColorPants_ColorBoots
+// Garment - Long Sleeves
+class CLASS(Unit_Garment_LS_BS_BP_BB): CLASS(Unit_Garment_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Garment_LS_BS_BP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_garment_bs_bp_bb_co.paa)};
+};
+class CLASS(Unit_Garment_LS_BS_GP_BB): CLASS(Unit_Garment_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Garment_LS_BS_GP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_garment_bs_gp_bb_co.paa)};
+};
+class CLASS(Unit_Garment_LS_OS_EP_TB): CLASS(Unit_Garment_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Garment_LS_OS_EP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_garment_os_ep_tb_co.paa)};
+};
+class CLASS(Unit_Garment_LS_GS_GP_BB): CLASS(Unit_Garment_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Garment_LS_GS_GP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_garment_gs_gp_bb_co.paa)};
+};
+class CLASS(Unit_Garment_LS_GS_BP_BB): CLASS(Unit_Garment_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Garment_LS_GS_BP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_garment_gs_bp_bb_co.paa)};
+};
+class CLASS(Unit_Garment_LS_GS_EP_TB): CLASS(Unit_Garment_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Garment_LS_GS_EP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_garment_gs_ep_tb_co.paa)};
+};
+class CLASS(Unit_Garment_LS_ES_EP_TB): CLASS(Unit_Garment_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Garment_LS_ES_EP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_garment_es_ep_tb_co.paa)};
+};
+class CLASS(Unit_Garment_LS_ES_BP_BB): CLASS(Unit_Garment_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Garment_LS_ES_BP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_garment_es_bp_bb_co.paa)};
+};
+class CLASS(Unit_Garment_LS_ES_GP_BB): CLASS(Unit_Garment_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Garment_LS_ES_GP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_garment_es_gp_bb_co.paa)};
+};
+class CLASS(Unit_Garment_LS_TS_TP_TB): CLASS(Unit_Garment_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Garment_LS_TS_TP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_garment_ts_tp_tb_co.paa)};
+};
+class CLASS(Unit_Garment_LS_GS_TP_TB): CLASS(Unit_Garment_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Garment_LS_GS_TP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_garment_gs_tp_tb_co.paa)};
+};
+class CLASS(Unit_Garment_LS_TS_GP_BB): CLASS(Unit_Garment_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Garment_LS_TS_GP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_garment_ts_gp_bb_co.paa)};
+};
+class CLASS(Unit_Garment_LS_OS_TP_TB): CLASS(Unit_Garment_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Garment_LS_OS_TP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_garment_os_tp_tb_co.paa)};
+};
+class CLASS(Unit_Garment_LS_TS_BP_BB): CLASS(Unit_Garment_LS_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Garment_LS_TS_BP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_garment_ts_bp_bb_co.paa)};
+};
