+++ addons/units/CfgVehicles/garment_rolled_item.hpp
@@ -1,0 +1,126 @@
+class CLASS(Item_Uniform_Garment_RS_BS_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_BS_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_BS_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Garment_RS_BS_GP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_BS_GP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_BS_GP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Garment_RS_OS_EP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_OS_EP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_OS_EP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Garment_RS_GS_GP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_GS_GP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_GS_GP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Garment_RS_GS_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_GS_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_GS_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Garment_RS_GS_EP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_GS_EP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_GS_EP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Garment_RS_ES_EP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_ES_EP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_ES_EP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Garment_RS_ES_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_ES_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_ES_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Garment_RS_ES_GP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_ES_GP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_ES_GP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Garment_RS_TS_TP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_TS_TP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_TS_TP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Garment_RS_GS_TP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_GS_TP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_GS_TP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Garment_RS_TS_GP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_TS_GP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_TS_GP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Garment_RS_OS_TP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_OS_TP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_OS_TP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Garment_RS_TS_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_TS_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_TS_BP_BB),1);
+    };
+};
