+++ addons/units/CfgVehicles/plaid_jeans_helpers.hpp
@@ -1,0 +1,58 @@
+// Naming pattern: SizeSleeves_ColorShirt_ColorPants_ColorBoots
+// Plaid with Jeans
+class CLASS(Unit_Plaid_JP_GY_BP_BB): CLASS(Unit_TShirt_Floral_Denim_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Plaid_JP_GY_BP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_plaid_denim_gy_bp_bb_co.paa)
+    };
+};
+class CLASS(Unit_Plaid_JP_LP_BP_BB): CLASS(Unit_TShirt_Floral_Denim_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Plaid_JP_LP_BP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_plaid_denim_lp_bp_bb_co.paa)
+    };
+};
+class CLASS(Unit_Plaid_BP_LPP_BP_BB): CLASS(Unit_TShirt_Floral_Denim_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Plaid_BP_LPP_BP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_plaid_denim_lpp_bp_bb_co.paa)
+    };
+};
+class CLASS(Unit_Plaid_BP_LY_BP_BB): CLASS(Unit_TShirt_Floral_Denim_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Plaid_BP_LY_BP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_plaid_denim_ly_bp_bb_co.paa)
+    };
+};
+class CLASS(Unit_Plaid_JP_CB_BP_BB): CLASS(Unit_TShirt_Floral_Denim_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Plaid_JP_CB_BP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_plaid_denim_cb_bp_bb_co.paa)
+    };
+};
+class CLASS(Unit_Plaid_JP_RO_BP_BB): CLASS(Unit_TShirt_Floral_Denim_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Plaid_JP_RO_BP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_plaid_denim_ro_bp_bb_co.paa)
+    };
+};
+class CLASS(Unit_Plaid_JP_RT_BP_BB): CLASS(Unit_TShirt_Floral_Denim_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Plaid_JP_RT_BP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_plaid_denim_rt_bp_bb_co.paa)
+    };
+};
