+++ addons/units/CfgVehicles/plaid_jeans_item.hpp
@@ -1,0 +1,63 @@
+class CLASS(Item_Uniform_Plaid_JP_GY_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_JP_GY_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Plaid_JP_GY_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Plaid_JP_LP_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, TyroneMF";
+    displayName = CSTRING(Uniform_Plaid_JP_LP_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Plaid_JP_LP_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Plaid_BP_LP_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_BP_LPP_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Plaid_BP_LP_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Plaid_BP_LY_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_BP_LY_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Plaid_BP_LY_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Plaid_JP_CB_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_JP_CB_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Plaid_JP_CB_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Plaid_JP_RO_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_JP_RO_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Plaid_JP_RO_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Plaid_JP_RT_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_JP_RT_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Plaid_JP_RT_BP_BB),1);
+    };
+};
