+++ addons/units/CfgVehicles/plaid_shorts_helpers.hpp
@@ -1,0 +1,58 @@
+// Naming pattern: SizeSleeves_ColorShirt_ColorPants_ColorBoots
+// Plaid with Shorts
+class CLASS(Unit_Plaid_SP_GY_BP_BB): CLASS(Unit_TShirt_Floral_Shorts_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Plaid_SP_GY_BP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_plaid_shorts_gy_bp_bb_co.paa)
+    };
+};
+class CLASS(Unit_Plaid_SP_LP_BP_BB): CLASS(Unit_TShirt_Floral_Shorts_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Plaid_SP_LP_BP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_plaid_shorts_lp_bp_bb_co.paa)
+    };
+};
+class CLASS(Unit_Plaid_SP_LPP_BP_BB): CLASS(Unit_TShirt_Floral_Shorts_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Plaid_SP_LPP_BP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_plaid_shorts_lpp_bp_bb_co.paa)
+    };
+};
+class CLASS(Unit_Plaid_SP_LY_BP_BB): CLASS(Unit_TShirt_Floral_Shorts_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Plaid_SP_LY_BP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_plaid_shorts_ly_bp_bb_co.paa)
+    };
+};
+class CLASS(Unit_Plaid_SP_CB_BP_BB): CLASS(Unit_TShirt_Floral_Shorts_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Plaid_SP_CB_BP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_plaid_shorts_cb_bp_bb_co.paa)
+    };
+};
+class CLASS(Unit_Plaid_SP_RO_BP_BB): CLASS(Unit_TShirt_Floral_Shorts_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Plaid_SP_RO_BP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_plaid_shorts_ro_bp_bb_co.paa)
+    };
+};
+class CLASS(Unit_Plaid_SP_RT_BP_BB): CLASS(Unit_TShirt_Floral_Shorts_Base) {
+    scope = 1;
+    author = "Trenchgun, Tyrone";
+    uniformClass = QCLASS(Uniform_Plaid_SP_RT_BP_BB);
+    hiddenSelectionsTextures[] = {
+        QPATHTOF(data\uniform_plaid_shorts_rt_bp_bb_co.paa)
+    };
+};
