+++ addons/units/CfgVehicles/plaid_shorts_item.hpp
@@ -1,0 +1,63 @@
+class CLASS(Item_Uniform_Plaid_SP_GY_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_SP_GY_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Plaid_SP_GY_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Plaid_SP_LP_BP_BB): Item_Base_F { // Pink
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, TyroneMF";
+    displayName = CSTRING(Uniform_Plaid_SP_LP_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Plaid_SP_LP_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Plaid_SP_LPP_BP_BB): Item_Base_F { // Purple
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_SP_LPP_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Plaid_SP_LPP_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Plaid_SP_LY_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_SP_LY_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Plaid_SP_LY_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Plaid_SP_CB_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_SP_CB_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Plaid_SP_CB_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Plaid_SP_RO_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_SP_RO_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Plaid_SP_RO_BP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Plaid_SP_RT_BP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_SP_RT_BP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Plaid_SP_RT_BP_BB),1);
+    };
+};
