+++ addons/units/CfgVehicles/polo_cord_helpers.hpp
@@ -1,0 +1,26 @@
+// Naming pattern: SizeSleeves_ColorShirt_ColorPants_ColorBoots
+// Polo - Cord Pants
+class CLASS(Unit_Polo_CP_LS_TP_OB): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Polo_CP_LS_TP_OB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_ls_tp_ob_co.paa)};
+};
+class CLASS(Unit_Polo_CP_RS_LP_BB): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas";
+    uniformClass = QCLASS(Uniform_Polo_CP_RS_LP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_rs_lp_bb_co.paa)};
+};
+class CLASS(Unit_Polo_CP_BS_TP_BB): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Rory";
+    uniformClass = QCLASS(Uniform_Polo_CP_BS_TP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_bs_tp_bb_co.paa)};
+};
+class CLASS(Unit_Polo_CP_BS_TP_BB_NoLogo): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Kresky";
+    uniformClass = QCLASS(Uniform_Polo_CP_BS_TP_BB_NoLogo);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_bs_tp_bb_nologo_co.paa)};
+};
