+++ addons/units/CfgVehicles/polo_cord_item.hpp
@@ -1,0 +1,36 @@
+class CLASS(Item_Uniform_Polo_CP_LS_TP_OB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_CP_LS_TP_OB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_CP_LS_TP_OB),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_CP_RS_LP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_CP_RS_LP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_CP_RS_LP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_CP_BS_TP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_CP_BS_TP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_CP_BS_TP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_CP_BS_TP_BB_NoLogo): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_CP_BS_TP_BB_NoLogo);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_CP_BS_TP_BB_NoLogo),1);
+    };
+};
