+++ addons/units/CfgVehicles/polo_tactical_helpers.hpp
@@ -1,0 +1,134 @@
+// Naming pattern: SizeSleeves_ColorShirt_ColorPants_ColorBoots
+// Polo - Tactical Pants
+class CLASS(Unit_Polo_TP_LS_TP_TB): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Rory";
+    uniformClass = QCLASS(Uniform_Polo_TP_LS_TP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_ls_tp_tb_co.paa)};
+};
+class CLASS(Unit_Polo_TP_LS_TP_TB_NoLogo): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Kresky";
+    uniformClass = QCLASS(Uniform_Polo_TP_LS_TP_TB_NoLogo);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_ls_tp_tb_nologo_co.paa)};
+};
+class CLASS(Unit_Polo_TP_TS_GP_BB): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Rory";
+    uniformClass = QCLASS(Uniform_Polo_TP_TS_GP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_ts_gp_bb_co.paa)};
+};
+class CLASS(Unit_Polo_TP_TS_GP_BB_NoLogo): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Kresky";
+    uniformClass = QCLASS(Uniform_Polo_TP_TS_GP_BB_NoLogo);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_ts_gp_bb_nologo_co.paa)};
+};
+class CLASS(Unit_Polo_TP_BS_TP_TB): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Rory";
+    uniformClass = QCLASS(Uniform_Polo_TP_BS_TP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_bs_tp_tb_co.paa)};
+};
+class CLASS(Unit_Polo_TP_BS_TP_TB_NoLogo): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Kresky";
+    uniformClass = QCLASS(Uniform_Polo_TP_BS_TP_TB_NoLogo);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_bs_tp_tb_nologo_co.paa)};
+};
+class CLASS(Unit_Polo_TP_BS_LP_BB): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Rory";
+    uniformClass = QCLASS(Uniform_Polo_TP_BS_LP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_bs_lp_bb_co.paa)};
+};
+class CLASS(Unit_Polo_TP_BS_LP_BB_NoLogo): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Kresky";
+    uniformClass = QCLASS(Uniform_Polo_TP_BS_LP_BB_NoLogo);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_bs_lp_bb_nologo_co.paa)};
+};
+class CLASS(Unit_Polo_TP_LS_GP_BB): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Rory";
+    uniformClass = QCLASS(Uniform_Polo_TP_LS_GP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_ls_gp_bb_co.paa)};
+};
+class CLASS(Unit_Polo_TP_LS_GP_BB_NoLogo): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Kresky";
+    uniformClass = QCLASS(Uniform_Polo_TP_LS_GP_BB_NoLogo);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_ls_gp_bb_nologo_co.paa)};
+};
+class CLASS(Unit_Polo_TP_OS_TP_BB): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Rory";
+    uniformClass = QCLASS(Uniform_Polo_TP_OS_TP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_os_tp_bb_co.paa)};
+};
+class CLASS(Unit_Polo_TP_OS_TP_BB_NoLogo): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Kresky";
+    uniformClass = QCLASS(Uniform_Polo_TP_OS_TP_BB_NoLogo);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_os_tp_bb_nologo_co.paa)};
+};
+class CLASS(Unit_Polo_TP_OS_LP_BB): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Rory";
+    uniformClass = QCLASS(Uniform_Polo_TP_OS_LP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_os_lp_bb_co.paa)};
+};
+class CLASS(Unit_Polo_TP_OS_LP_BB_NoLogo): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Kresky";
+    uniformClass = QCLASS(Uniform_Polo_TP_OS_LP_BB_NoLogo);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_os_lp_bb_nologo_co.paa)};
+};
+class CLASS(Unit_Polo_TP_GS_TP_TB): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Rory";
+    uniformClass = QCLASS(Uniform_Polo_TP_GS_TP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_gs_tp_tb_co.paa)};
+};
+class CLASS(Unit_Polo_TP_GS_TP_TB_NoLogo): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Kresky";
+    uniformClass = QCLASS(Uniform_Polo_TP_GS_TP_TB_NoLogo);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_gs_tp_tb_nologo_co.paa)};
+};
+class CLASS(Unit_Polo_TP_WS_TP_TB): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Rory";
+    uniformClass = QCLASS(Uniform_Polo_TP_WS_TP_TB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_ws_tp_tb_co.paa)};
+};
+class CLASS(Unit_Polo_TP_WS_TP_TB_NoLogo): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Kresky";
+    uniformClass = QCLASS(Uniform_Polo_TP_WS_TP_TB_NoLogo);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_ws_tp_tb_nologo_co.paa)};
+};
+class CLASS(Unit_Polo_WS_LP_BB): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Rory";
+    uniformClass = QCLASS(Uniform_Polo_TP_WS_LP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_ws_lp_bb_co.paa)};
+};
+class CLASS(Unit_Polo_WS_LP_BB_NoLogo): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Kresky";
+    uniformClass = QCLASS(Uniform_Polo_TP_WS_LP_BB_NoLogo);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_ws_lp_bb_nologo_co.paa)};
+};
+class CLASS(Unit_Polo_TP_WS_GP_BB): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Rory";
+    uniformClass = QCLASS(Uniform_Polo_TP_WS_GP_BB);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_ws_gp_bb_co.paa)};
+};
+class CLASS(Unit_Polo_TP_WS_GP_BB_NoLogo): CLASS(Unit_Polo_TP_Base) {
+    scope = 1;
+    author = "Pomigit, Jonpas, Kresky";
+    uniformClass = QCLASS(Uniform_Polo_TP_WS_GP_BB_NoLogo);
+    hiddenSelectionsTextures[] = {QPATHTOF(data\uniform_polo_ws_gp_bb_nologo_co.paa)};
+};
