+++ addons/units/CfgVehicles/polo_tactical_item.hpp
@@ -1,0 +1,198 @@
+class CLASS(Item_Uniform_Polo_TP_LS_TP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_LS_TP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_LS_TP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_LS_TP_TB_NoLogo): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_LS_TP_TB_NoLogo);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_LS_TP_TB_NoLogo),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_TS_GP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_TS_GP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_TS_GP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_TS_GP_BB_NoLogo): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_TS_GP_BB_NoLogo);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_TS_GP_BB_NoLogo),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_BS_TP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_BS_TP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_BS_TP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_BS_TP_TB_NoLogo): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_BS_TP_TB_NoLogo);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_BS_TP_TB_NoLogo),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_BS_LP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_BS_LP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_BS_LP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_BS_LP_BB_NoLogo): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_BS_LP_BB_NoLogo);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_BS_LP_BB_NoLogo),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_LS_GP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_LS_GP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_LS_GP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_LS_GP_BB_NoLogo): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_LS_GP_BB_NoLogo);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_LS_GP_BB_NoLogo),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_OS_TP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_OS_TP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_OS_TP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_OS_TP_BB_NoLogo): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_OS_TP_BB_NoLogo);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_OS_TP_BB_NoLogo),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_OS_LP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_OS_LP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_OS_LP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_OS_LP_BB_NoLogo): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_OS_LP_BB_NoLogo);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_OS_LP_BB_NoLogo),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_GS_TP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_GS_TP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_GS_TP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_GS_TP_TB_NoLogo): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_GS_TP_TB_NoLogo);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_GS_TP_TB_NoLogo),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_WS_TP_TB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_WS_TP_TB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_WS_TP_TB),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_WS_TP_TB_NoLogo): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_WS_TP_TB_NoLogo);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_WS_TP_TB_NoLogo),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_WS_LP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_WS_LP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_WS_LP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_WS_LP_BB_NoLogo): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_WS_LP_BB_NoLogo);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_WS_LP_BB_NoLogo),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_WS_GP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_WS_GP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_WS_GP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_Polo_TP_WS_GP_BB_NoLogo): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_Polo_TP_WS_GP_BB_NoLogo);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_WS_GP_BB_NoLogo),1);
+    };
+};
