+++ addons/units/CfgVehicles/tshirt_jeans_item.hpp
@@ -1,0 +1,54 @@
+class CLASS(Item_Uniform_TShirt_JP_GS_LP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_TShirt_JP_GS_LP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_TShirt_JP_GS_LP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_TShirt_JP_GS_TP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_TShirt_JP_GS_TP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_TShirt_JP_GS_TP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_TShirt_JP_BS_LP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_TShirt_JP_BS_LP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_TShirt_JP_BS_LP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_TShirt_JP_BS_TP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_TShirt_JP_BS_TP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_TShirt_JP_BS_TP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_TShirt_JP_LS_LP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_TShirt_JP_LS_LP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_TShirt_JP_LS_LP_BB),1);
+    };
+};
+class CLASS(Item_Uniform_TShirt_JP_WS_LP_BB): Item_Base_F {
+    MACRO_ITEM_COMMON
+    author = "Jonpas";
+    displayName = CSTRING(Uniform_TShirt_JP_WS_LP_BB);
+
+    class TransportItems {
+        MACRO_ADDITEM(CLASS(Uniform_TShirt_JP_WS_LP_BB),1);
+    };
+};
