--- addons/units/CfgVehicles_GroundItems.hpp
@@ -1,1253 +1,0 @@
-// Included in CfgVehicles.hpp
-
-#define MACRO_ITEM_COMMON \
-    dlc = QUOTE(PREFIX); \
-    scope = 2; \
-    scopeCurator = 2; \
-    vehicleClass = "ItemsUniforms"; \
-    model = "\A3\Weapons_F\DummySuitpack.p3d";
-
-class Item_Base_F;
-class CLASS(Item_Uniform_Combat_LS_ATACS_ATX): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Tibernius, Tyrone";
-    displayName = CSTRING(Uniform_Combat_LS_ATACS_ATX);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_ATACS_ATX),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_ATACS_AUX): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Tibernius, Tyrone";
-    displayName = CSTRING(Uniform_Combat_LS_ATACS_AUX);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_ATACS_AUX),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_ATACS_IX): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Tibernius, Tyrone";
-    displayName = CSTRING(Uniform_Combat_LS_ATACS_IX);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_ATACS_IX),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_ATACS_LEX): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Tibernius, Tyrone";
-    displayName = CSTRING(Uniform_Combat_LS_ATACS_LEX);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_ATACS_LEX),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_ATACS_ATX): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Tibernius, Tyrone";
-    displayName = CSTRING(Uniform_Combat_RS_ATACS_ATX);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_ATACS_ATX),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_ATACS_AUX): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Tibernius, Tyrone";
-    displayName = CSTRING(Uniform_Combat_RS_ATACS_AUX);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_ATACS_AUX),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_ATACS_IX): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Tibernius, Tyrone";
-    displayName = CSTRING(Uniform_Combat_RS_ATACS_IX);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_ATACS_IX),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_ATACS_LEX): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Tibernius, Tyrone";
-    displayName = CSTRING(Uniform_Combat_RS_ATACS_LEX);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_ATACS_LEX),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_BS_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Kresky";
-    displayName = CSTRING(Uniform_Combat_LS_BS_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_BS_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_BS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_LS_BS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_BS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_BS_GP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_LS_BS_GP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_BS_GP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_BS_TP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_LS_BS_TP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_BS_TP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_BS_TP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_LS_BS_TP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_BS_TP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_GS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_LS_GS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_GS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_GS_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_LS_GS_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_GS_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_GS_TP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_LS_GS_TP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_GS_TP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_BS_DGP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_LS_BS_DGP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_BS_DGP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_TS_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_LS_TS_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_TS_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_TS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_LS_TS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_TS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_TS_GP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_LS_TS_GP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_TS_GP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_TS_TP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_LS_TS_TP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_TS_TP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_CBPS_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "GilleeDoo, Tyrone";
-    displayName = CSTRING(Uniform_Combat_LS_CBPS_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CBPS_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_CDES_TP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "GilleeDoo, Tyrone";
-    displayName = CSTRING(Uniform_Combat_LS_CDES_TP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CDES_TP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_CDFS_GP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "GilleeDoo, Tyrone";
-    displayName = CSTRING(Uniform_Combat_LS_CDFS_GP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CDFS_GP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_CLES_TP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "GilleeDoo, Tyrone";
-    displayName = CSTRING(Uniform_Combat_LS_CLES_TP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CLES_TP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_CLFS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "GilleeDoo, Tyrone";
-    displayName = CSTRING(Uniform_Combat_LS_CLFS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CLFS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_CDBS_GP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_LS_CDBS_GP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CDBS_GP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_CLBS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_LS_CLBS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CLBS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_CLRS_TP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_LS_CLRS_TP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CLRS_TP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_CPS_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_LS_CPS_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_CPS_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_LS_WS_WP_WB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas, Kresky";
-    displayName = CSTRING(Uniform_Combat_LS_WS_WP_WB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_LS_WS_WP_WB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_BS_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas, Kresky";
-    displayName = CSTRING(Uniform_Combat_RS_BS_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_BS_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_BS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_RS_BS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_BS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_BS_GP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_RS_BS_GP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_BS_GP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_BS_TP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_RS_BS_TP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_BS_TP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_BS_TP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_RS_BS_TP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_BS_TP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_GS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_RS_GS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_GS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_GS_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_RS_GS_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_GS_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_GS_TP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_RS_GS_TP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_GS_TP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_BS_DGP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_RS_BS_DGP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_BS_DGP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_TS_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_RS_TS_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_TS_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_TS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_RS_TS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_TS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_TS_GP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_RS_TS_GP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_TS_GP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_TS_TP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_RS_TS_TP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_TS_TP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_CBPS_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "GilleeDoo, Tyrone";
-    displayName = CSTRING(Uniform_Combat_RS_CBPS_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_CBPS_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_CDES_TP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "GilleeDoo, Tyrone";
-    displayName = CSTRING(Uniform_Combat_RS_CDES_TP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_CDES_TP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_CDFS_GP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "GilleeDoo, Tyrone";
-    displayName = CSTRING(Uniform_Combat_RS_CDFS_GP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_CDFS_GP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_CLES_TP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "GilleeDoo, Tyrone";
-    displayName = CSTRING(Uniform_Combat_RS_CLES_TP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_CLES_TP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_CLFS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "GilleeDoo, Tyrone";
-    displayName = CSTRING(Uniform_Combat_RS_CLFS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_CLFS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_CDBS_GP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_RS_CDBS_GP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_CDBS_GP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_CLBS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_RS_CLBS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_CLBS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_CLRS_TP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_RS_CLRS_TP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_CLRS_TP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Combat_RS_CPS_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Combat_RS_CPS_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Combat_RS_CPS_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_LS_BS_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_LS_BS_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_LS_BS_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_LS_BS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_LS_BS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_LS_BS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_LS_OS_EP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_LS_OS_EP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_LS_OS_EP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_LS_GS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_LS_GS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_LS_GS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_LS_GS_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_LS_GS_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_LS_GS_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_LS_GS_EP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_LS_GS_EP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_LS_GS_EP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_LS_ES_EP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_LS_ES_EP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_LS_ES_EP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_LS_ES_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_LS_ES_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_LS_ES_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_LS_ES_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_LS_ES_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_LS_ES_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_LS_TS_TP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_LS_TS_TP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_LS_TS_TP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_LS_GS_TP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_LS_GS_TP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_LS_GS_TP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_LS_TS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_LS_TS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_LS_TS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_LS_OS_TP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_LS_OS_TP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_LS_OS_TP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_LS_TS_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_LS_TS_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_LS_TS_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_RS_BS_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_RS_BS_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_BS_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_RS_BS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_RS_BS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_BS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_RS_OS_EP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_RS_OS_EP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_OS_EP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_RS_GS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_RS_GS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_GS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_RS_GS_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_RS_GS_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_GS_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_RS_GS_EP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_RS_GS_EP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_GS_EP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_RS_ES_EP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_RS_ES_EP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_ES_EP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_RS_ES_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_RS_ES_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_ES_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_RS_ES_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_RS_ES_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_ES_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_RS_TS_TP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_RS_TS_TP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_TS_TP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_RS_GS_TP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_RS_GS_TP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_GS_TP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_RS_TS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_RS_TS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_TS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_RS_OS_TP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_RS_OS_TP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_OS_TP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Garment_RS_TS_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Garment_RS_TS_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Garment_RS_TS_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_LS_TP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_LS_TP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_LS_TP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_LS_TP_TB_NoLogo): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_LS_TP_TB_NoLogo);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_LS_TP_TB_NoLogo),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_TS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_TS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_TS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_TS_GP_BB_NoLogo): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_TS_GP_BB_NoLogo);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_TS_GP_BB_NoLogo),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_BS_TP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_BS_TP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_BS_TP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_BS_TP_TB_NoLogo): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_BS_TP_TB_NoLogo);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_BS_TP_TB_NoLogo),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_BS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_BS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_BS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_BS_LP_BB_NoLogo): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_BS_LP_BB_NoLogo);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_BS_LP_BB_NoLogo),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_LS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_LS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_LS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_LS_GP_BB_NoLogo): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_LS_GP_BB_NoLogo);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_LS_GP_BB_NoLogo),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_OS_TP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_OS_TP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_OS_TP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_OS_TP_BB_NoLogo): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_OS_TP_BB_NoLogo);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_OS_TP_BB_NoLogo),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_OS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_OS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_OS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_OS_LP_BB_NoLogo): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_OS_LP_BB_NoLogo);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_OS_LP_BB_NoLogo),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_GS_TP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_GS_TP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_GS_TP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_GS_TP_TB_NoLogo): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_GS_TP_TB_NoLogo);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_GS_TP_TB_NoLogo),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_WS_TP_TB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_WS_TP_TB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_WS_TP_TB),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_WS_TP_TB_NoLogo): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_WS_TP_TB_NoLogo);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_WS_TP_TB_NoLogo),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_WS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_WS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_WS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_WS_LP_BB_NoLogo): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_WS_LP_BB_NoLogo);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_WS_LP_BB_NoLogo),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_WS_GP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_WS_GP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_WS_GP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_TP_WS_GP_BB_NoLogo): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_TP_WS_GP_BB_NoLogo);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_TP_WS_GP_BB_NoLogo),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_CP_LS_TP_OB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_CP_LS_TP_OB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_CP_LS_TP_OB),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_CP_RS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_CP_RS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_CP_RS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_CP_BS_TP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_CP_BS_TP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_CP_BS_TP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Polo_CP_BS_TP_BB_NoLogo): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_Polo_CP_BS_TP_BB_NoLogo);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Polo_CP_BS_TP_BB_NoLogo),1);
-    };
-};
-class CLASS(Item_Uniform_TShirt_JP_GS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_TShirt_JP_GS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_TShirt_JP_GS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_TShirt_JP_GS_TP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_TShirt_JP_GS_TP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_TShirt_JP_GS_TP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_TShirt_JP_BS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_TShirt_JP_BS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_TShirt_JP_BS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_TShirt_JP_BS_TP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_TShirt_JP_BS_TP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_TShirt_JP_BS_TP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_TShirt_JP_LS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_TShirt_JP_LS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_TShirt_JP_LS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_TShirt_JP_WS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Uniform_TShirt_JP_WS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_TShirt_JP_WS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Floral_JP_BS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Floral_JP_BS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Floral_JP_BS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Floral_JP_RS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Floral_JP_RS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Floral_JP_RS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Floral_JP_GS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Floral_JP_GS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Floral_JP_GS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Floral_JP_WS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Floral_JP_WS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Floral_JP_WS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Floral_JP_BYS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Floral_JP_BYS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Floral_JP_BYS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Floral_SP_BS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Floral_SP_BS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Floral_SP_BS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Floral_SP_RS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Floral_SP_RS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Floral_SP_RS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Floral_SP_GS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Floral_SP_GS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Floral_SP_GS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Floral_SP_WS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Floral_SP_WS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Floral_SP_WS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Floral_SP_BYS_LP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Floral_SP_BYS_LP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Floral_SP_BYS_LP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Plaid_JP_GY_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Plaid_JP_GY_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Plaid_JP_GY_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Plaid_JP_LP_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, TyroneMF";
-    displayName = CSTRING(Uniform_Plaid_JP_LP_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Plaid_JP_LP_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Plaid_BP_LP_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Plaid_BP_LPP_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Plaid_BP_LP_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Plaid_BP_LY_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Plaid_BP_LY_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Plaid_BP_LY_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Plaid_JP_CB_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Plaid_JP_CB_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Plaid_JP_CB_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Plaid_JP_RO_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Plaid_JP_RO_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Plaid_JP_RO_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Plaid_JP_RT_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Plaid_JP_RT_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Plaid_JP_RT_BP_BB),1);
-    };
-};
-// Shorts
-class CLASS(Item_Uniform_Plaid_SP_GY_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Plaid_SP_GY_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Plaid_SP_GY_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Plaid_SP_LP_BP_BB): Item_Base_F { // Pink
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, TyroneMF";
-    displayName = CSTRING(Uniform_Plaid_SP_LP_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Plaid_SP_LP_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Plaid_SP_LPP_BP_BB): Item_Base_F { // Purple
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Plaid_SP_LPP_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Plaid_SP_LPP_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Plaid_SP_LY_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Plaid_SP_LY_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Plaid_SP_LY_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Plaid_SP_CB_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Plaid_SP_CB_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Plaid_SP_CB_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Plaid_SP_RO_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Plaid_SP_RO_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Plaid_SP_RO_BP_BB),1);
-    };
-};
-class CLASS(Item_Uniform_Plaid_SP_RT_BP_BB): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Trenchgun, Tyrone";
-    displayName = CSTRING(Uniform_Plaid_SP_RT_BP_BB);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Uniform_Plaid_SP_RT_BP_BB),1);
-    };
-};
-class CLASS(Item_Suit_VIP): Item_Base_F {
-    MACRO_ITEM_COMMON
-    author = "Jonpas";
-    displayName = CSTRING(Suit_VIP);
-
-    class TransportItems {
-        MACRO_ADDITEM(CLASS(Suit_VIP),1);
-    };
-};
