+++ addons/units/CfgWeapons/combat_long_plaid_uniforms.hpp
@@ -1,0 +1,92 @@
+// Naming pattern: TypeSleevesOrPants_ColorShirt_ColorPants_ColorBoots
+// Plaid - Long sleeves
+class CLASS(Uniform_Combat_LS_CBPS_BP_BB): CLASS(Uniform_Base) { // CBP = Buffalo Plaid
+    scope = 2;
+    author = "GilleeDoo, Tyrone";
+    displayName = CSTRING(Uniform_Combat_LS_CBPS_BP_BB);
+    picture = QPATHTOF(UI\uniform_combat_ls_cbps_bp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_CBPS_BP_BB);
+    };
+};
+class CLASS(Uniform_Combat_LS_CDES_TP_TB): CLASS(Uniform_Base) { // CDE = Plaid Dark Earth
+    scope = 2;
+    author = "GilleeDoo, Tyrone";
+    displayName = CSTRING(Uniform_Combat_LS_CDES_TP_TB);
+    picture = QPATHTOF(UI\uniform_combat_ls_cdes_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_CDES_TP_TB);
+    };
+};
+class CLASS(Uniform_Combat_LS_CDFS_GP_TB): CLASS(Uniform_Base) { // CDF = Plaid Dark Foliage
+    scope = 2;
+    author = "GilleeDoo, Tyrone";
+    displayName = CSTRING(Uniform_Combat_LS_CDFS_GP_TB);
+    picture = QPATHTOF(UI\uniform_combat_ls_cdfs_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_CDFS_GP_TB);
+    };
+};
+class CLASS(Uniform_Combat_LS_CLES_TP_BB): CLASS(Uniform_Base) { // CLE = Plaid Light Earth
+    scope = 2;
+    author = "GilleeDoo, Tyrone";
+    displayName = CSTRING(Uniform_Combat_LS_CLES_TP_BB);
+    picture = QPATHTOF(UI\uniform_combat_ls_cles_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_CLES_TP_BB);
+    };
+};
+class CLASS(Uniform_Combat_LS_CLFS_GP_BB): CLASS(Uniform_Base) { // CLF = Plaid Light Foliage
+    scope = 2;
+    author = "GilleeDoo, Tyrone";
+    displayName = CSTRING(Uniform_Combat_LS_CLFS_GP_BB);
+    picture = QPATHTOF(UI\uniform_combat_ls_clfs_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_CLFS_GP_BB);
+    };
+};
+class CLASS(Uniform_Combat_LS_CDBS_GP_TB): CLASS(Uniform_Base) { // CDB = Plaid Dark Blue
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_CDBS_GP_TB);
+    picture = QPATHTOF(UI\uniform_combat_ls_cdbs_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_CDBS_GP_TB);
+    };
+};
+class CLASS(Uniform_Combat_LS_CLBS_GP_BB): CLASS(Uniform_Base) { // CLB = Plaid Light Blue
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_CLBS_GP_BB);
+    picture = QPATHTOF(UI\uniform_combat_ls_clbs_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_CLBS_GP_BB);
+    };
+};
+class CLASS(Uniform_Combat_LS_CLRS_TP_BB): CLASS(Uniform_Base) { // CLR = Plaid Light Red
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_CLRS_TP_BB);
+    picture = QPATHTOF(UI\uniform_combat_ls_clrs_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_CLRS_TP_BB);
+    };
+};
+class CLASS(Uniform_Combat_LS_CPS_BP_BB): CLASS(Uniform_Base) { // CP = Plaid Purple
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_CPS_BP_BB);
+    picture = QPATHTOF(UI\uniform_combat_ls_cps_bp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_CPS_BP_BB);
+    };
+};
