+++ addons/units/CfgWeapons/combat_long_uniforms.hpp
@@ -1,0 +1,142 @@
+// Naming pattern: TypeSleevesOrPants_ColorShirt_ColorPants_ColorBoots
+// Combat - Long Sleeves
+class CLASS(Uniform_Combat_LS_BS_BP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Kresky";
+    displayName = CSTRING(Uniform_Combat_LS_BS_BP_BB);
+    picture = QPATHTOF(UI\uniform_combat_ls_bs_bp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_BS_BP_BB);
+    };
+};
+class CLASS(Uniform_Combat_LS_BS_GP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_BS_GP_BB);
+    picture = QPATHTOF(UI\uniform_combat_ls_bs_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_BS_GP_BB);
+    };
+};
+class CLASS(Uniform_Combat_LS_BS_GP_TB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_BS_GP_TB);
+    picture = QPATHTOF(UI\uniform_combat_ls_bs_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_BS_GP_TB);
+    };
+};
+class CLASS(Uniform_Combat_LS_BS_TP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_BS_TP_BB);
+    picture = QPATHTOF(UI\uniform_combat_ls_bs_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_BS_TP_BB);
+    };
+};
+class CLASS(Uniform_Combat_LS_BS_TP_TB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_BS_TP_TB);
+    picture = QPATHTOF(UI\uniform_combat_ls_bs_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_BS_TP_TB);
+    };
+};
+class CLASS(Uniform_Combat_LS_GS_GP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Kresky";
+    displayName = CSTRING(Uniform_Combat_LS_GS_GP_BB);
+    picture = QPATHTOF(UI\uniform_combat_ls_gs_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_GS_GP_BB);
+    };
+};
+class CLASS(Uniform_Combat_LS_GS_BP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_GS_BP_BB);
+    picture = QPATHTOF(UI\uniform_combat_ls_gs_bp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_GS_BP_BB);
+    };
+};
+class CLASS(Uniform_Combat_LS_GS_TP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_GS_TP_BB);
+    picture = QPATHTOF(UI\uniform_combat_ls_gs_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_GS_TP_BB);
+    };
+};
+class CLASS(Uniform_Combat_LS_BS_DGP_BB): CLASS(Uniform_Base) { // DG = Digital Green
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_BS_DGP_BB);
+    picture = QPATHTOF(UI\uniform_combat_ls_bs_dgp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_BS_DGP_BB);
+    };
+};
+class CLASS(Uniform_Combat_LS_TS_BP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_TS_BP_BB);
+    picture = QPATHTOF(UI\uniform_combat_ls_ts_bp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_TS_BP_BB);
+    };
+};
+class CLASS(Uniform_Combat_LS_TS_GP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_TS_GP_BB);
+    picture = QPATHTOF(UI\uniform_combat_ls_ts_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_TS_GP_BB);
+    };
+};
+class CLASS(Uniform_Combat_LS_TS_GP_TB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_LS_TS_GP_TB);
+    picture = QPATHTOF(UI\uniform_combat_ls_ts_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_TS_GP_TB);
+    };
+};
+class CLASS(Uniform_Combat_LS_TS_TP_TB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Kresky";
+    displayName = CSTRING(Uniform_Combat_LS_TS_TP_TB);
+    picture = QPATHTOF(UI\uniform_combat_ls_ts_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_TS_TP_TB);
+    };
+};
+class CLASS(Uniform_Combat_LS_WS_WP_WB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Kresky";
+    displayName = CSTRING(Uniform_Combat_LS_WS_WP_WB);
+    picture = QPATHTOF(UI\uniform_combat_ls_ws_wp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_LS_WS_WP_WB);
+    };
+};
