+++ addons/units/CfgWeapons/combat_rolled_plaid_uniforms.hpp
@@ -1,0 +1,92 @@
+// Naming pattern: TypeSleevesOrPants_ColorShirt_ColorPants_ColorBoots
+// Plaid - Rolled sleeves
+class CLASS(Uniform_Combat_RS_CBPS_BP_BB): CLASS(Uniform_Base) { // CBP = Buffalo Plaid
+    scope = 2;
+    author = "GilleeDoo, Tyrone";
+    displayName = CSTRING(Uniform_Combat_RS_CBPS_BP_BB);
+    picture = QPATHTOF(UI\uniform_combat_rs_cbps_bp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_CBPS_BP_BB);
+    };
+};
+class CLASS(Uniform_Combat_RS_CDES_TP_TB): CLASS(Uniform_Base) { // CDE = Plaid Dark Earth
+    scope = 2;
+    author = "GilleeDoo, Tyrone";
+    displayName = CSTRING(Uniform_Combat_RS_CDES_TP_TB);
+    picture = QPATHTOF(UI\uniform_combat_rs_cdes_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_CDES_TP_TB);
+    };
+};
+class CLASS(Uniform_Combat_RS_CDFS_GP_TB): CLASS(Uniform_Base) { // CDF = Plaid Dark Foliage
+    scope = 2;
+    author = "GilleeDoo, Tyrone";
+    displayName = CSTRING(Uniform_Combat_RS_CDFS_GP_TB);
+    picture = QPATHTOF(UI\uniform_combat_rs_cdfs_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_CDFS_GP_TB);
+    };
+};
+class CLASS(Uniform_Combat_RS_CLES_TP_BB): CLASS(Uniform_Base) { // CLE = Plaid Light Earth
+    scope = 2;
+    author = "GilleeDoo, Tyrone";
+    displayName = CSTRING(Uniform_Combat_RS_CLES_TP_BB);
+    picture = QPATHTOF(UI\uniform_combat_rs_cles_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_CLES_TP_BB);
+    };
+};
+class CLASS(Uniform_Combat_RS_CLFS_GP_BB): CLASS(Uniform_Base) { // CLF = Plaid Light Foliage
+    scope = 2;
+    author = "GilleeDoo, Tyrone";
+    displayName = CSTRING(Uniform_Combat_RS_CLFS_GP_BB);
+    picture = QPATHTOF(UI\uniform_combat_rs_clfs_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_CLFS_GP_BB);
+    };
+};
+class CLASS(Uniform_Combat_RS_CDBS_GP_TB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_CDBS_GP_TB);
+    picture = QPATHTOF(UI\uniform_combat_rs_cdbs_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_CDBS_GP_TB);
+    };
+};
+class CLASS(Uniform_Combat_RS_CLBS_GP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_CLBS_GP_BB);
+    picture = QPATHTOF(UI\uniform_combat_rs_clbs_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_CLBS_GP_BB);
+    };
+};
+class CLASS(Uniform_Combat_RS_CLRS_TP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_CLRS_TP_BB);
+    picture = QPATHTOF(UI\uniform_combat_rs_clrs_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_CLRS_TP_BB);
+    };
+};
+class CLASS(Uniform_Combat_RS_CPS_BP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_CPS_BP_BB);
+    picture = QPATHTOF(UI\uniform_combat_rs_cps_bp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_CPS_BP_BB);
+    };
+};
