+++ addons/units/CfgWeapons/combat_rolled_uniforms.hpp
@@ -1,0 +1,132 @@
+// Naming pattern: TypeSleevesOrPants_ColorShirt_ColorPants_ColorBoots
+// Combat - Rolled Sleeves
+class CLASS(Uniform_Combat_RS_BS_BP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Kresky";
+    displayName = CSTRING(Uniform_Combat_RS_BS_BP_BB);
+    picture = QPATHTOF(UI\uniform_combat_rs_bs_bp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_BS_BP_BB);
+    };
+};
+class CLASS(Uniform_Combat_RS_BS_GP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_BS_GP_BB);
+    picture = QPATHTOF(UI\uniform_combat_rs_bs_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_BS_GP_BB);
+    };
+};
+class CLASS(Uniform_Combat_RS_BS_GP_TB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_BS_GP_TB);
+    picture = QPATHTOF(UI\uniform_combat_rs_bs_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_BS_GP_TB);
+    };
+};
+class CLASS(Uniform_Combat_RS_BS_TP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_BS_TP_BB);
+    picture = QPATHTOF(UI\uniform_combat_rs_bs_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_BS_TP_BB);
+    };
+};
+class CLASS(Uniform_Combat_RS_BS_TP_TB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_BS_TP_TB);
+    picture = QPATHTOF(UI\uniform_combat_rs_bs_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_BS_TP_TB);
+    };
+};
+class CLASS(Uniform_Combat_RS_GS_GP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Kresky";
+    displayName = CSTRING(Uniform_Combat_RS_GS_GP_BB);
+    picture = QPATHTOF(UI\uniform_combat_rs_gs_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_GS_GP_BB);
+    };
+};
+class CLASS(Uniform_Combat_RS_GS_BP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_GS_BP_BB);
+    picture = QPATHTOF(UI\uniform_combat_rs_gs_bp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_GS_BP_BB);
+    };
+};
+class CLASS(Uniform_Combat_RS_GS_TP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_GS_TP_BB);
+    picture = QPATHTOF(UI\uniform_combat_rs_gs_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_GS_TP_BB);
+    };
+};
+class CLASS(Uniform_Combat_RS_BS_DGP_BB): CLASS(Uniform_Base) { // DG = Digital Green
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_BS_DGP_BB);
+    picture = QPATHTOF(UI\uniform_combat_rs_bs_dgp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_BS_DGP_BB);
+    };
+};
+class CLASS(Uniform_Combat_RS_TS_BP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_TS_BP_BB);
+    picture = QPATHTOF(UI\uniform_combat_rs_ts_bp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_TS_BP_BB);
+    };
+};
+class CLASS(Uniform_Combat_RS_TS_GP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_TS_GP_BB);
+    picture = QPATHTOF(UI\uniform_combat_rs_ts_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_TS_GP_BB);
+    };
+};
+class CLASS(Uniform_Combat_RS_TS_GP_TB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Combat_RS_TS_GP_TB);
+    picture = QPATHTOF(UI\uniform_combat_rs_ts_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_TS_GP_TB);
+    };
+};
+class CLASS(Uniform_Combat_RS_TS_TP_TB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Kresky";
+    displayName = CSTRING(Uniform_Combat_RS_TS_TP_TB);
+    picture = QPATHTOF(UI\uniform_combat_rs_ts_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Combat_RS_TS_TP_TB);
+    };
+};
