+++ addons/units/CfgWeapons/floral_jeans_uniforms.hpp
@@ -1,0 +1,52 @@
+// Naming pattern: TypeSleevesOrPants_ColorShirt_ColorPants_ColorBoots
+// Floral Shirts with Jeans
+class CLASS(Uniform_Floral_JP_BS_LP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Floral_JP_BS_LP_BB);
+    picture = QPATHTOF(UI\uniform_floral_jp_bs_lp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Floral_JP_BS_LP_BB);
+    };
+};
+class CLASS(Uniform_Floral_JP_RS_LP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Floral_JP_RS_LP_BB);
+    picture = QPATHTOF(UI\uniform_floral_jp_rs_lp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Floral_JP_RS_LP_BB);
+    };
+};
+class CLASS(Uniform_Floral_JP_GS_LP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Floral_JP_GS_LP_BB);
+    picture = QPATHTOF(UI\uniform_floral_jp_gs_lp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Floral_JP_GS_LP_BB);
+    };
+};
+class CLASS(Uniform_Floral_JP_WS_LP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Floral_JP_WS_LP_BB);
+    picture = QPATHTOF(UI\uniform_floral_jp_ws_lp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Floral_JP_WS_LP_BB);
+    };
+};
+class CLASS(Uniform_Floral_JP_BYS_LP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Floral_JP_BYS_LP_BB);
+    picture = QPATHTOF(UI\uniform_floral_jp_bys_lp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Floral_JP_BYS_LP_BB);
+    };
+};
