+++ addons/units/CfgWeapons/floral_shorts_uniforms.hpp
@@ -1,0 +1,52 @@
+// Naming pattern: TypeSleevesOrPants_ColorShirt_ColorPants_ColorBoots
+// Floral Shirts with Shorts
+class CLASS(Uniform_Floral_SP_BS_LP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Floral_SP_BS_LP_BB);
+    picture = QPATHTOF(UI\uniform_floral_sp_bs_lp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Floral_SP_BS_LP_BB);
+    };
+};
+class CLASS(Uniform_Floral_SP_RS_LP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Floral_SP_RS_LP_BB);
+    picture = QPATHTOF(UI\uniform_floral_sp_rs_lp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Floral_SP_RS_LP_BB);
+    };
+};
+class CLASS(Uniform_Floral_SP_GS_LP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Floral_SP_GS_LP_BB);
+    picture = QPATHTOF(UI\uniform_floral_sp_gs_lp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Floral_SP_GS_LP_BB);
+    };
+};
+class CLASS(Uniform_Floral_SP_WS_LP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Floral_SP_WS_LP_BB);
+    picture = QPATHTOF(UI\uniform_floral_sp_ws_lp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Floral_SP_WS_LP_BB);
+    };
+};
+class CLASS(Uniform_Floral_SP_BYS_LP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Floral_SP_BYS_LP_BB);
+    picture = QPATHTOF(UI\uniform_floral_sp_bys_lp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Floral_SP_BYS_LP_BB);
+    };
+};
