+++ addons/units/CfgWeapons/garment_rolled_uniforms.hpp
@@ -1,0 +1,142 @@
+// Naming pattern: TypeSleevesOrPants_ColorShirt_ColorPants_ColorBoots
+// Garment - Rolled Sleeves
+class CLASS(Uniform_Garment_RS_BS_BP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_BS_BP_BB);
+    picture = QPATHTOF(UI\uniform_garment_rs_bs_bp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Garment_RS_BS_BP_BB);
+    };
+};
+class CLASS(Uniform_Garment_RS_BS_GP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_BS_GP_BB);
+    picture = QPATHTOF(UI\uniform_garment_rs_bs_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Garment_RS_BS_GP_BB);
+    };
+};
+class CLASS(Uniform_Garment_RS_OS_EP_TB): CLASS(Uniform_Base) { // O = Brown, E = Grey
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_OS_EP_TB);
+    picture = QPATHTOF(UI\uniform_garment_rs_os_ep_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Garment_RS_OS_EP_TB);
+    };
+};
+class CLASS(Uniform_Garment_RS_GS_GP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_GS_GP_BB);
+    picture = QPATHTOF(UI\uniform_garment_rs_gs_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Garment_RS_GS_GP_BB);
+    };
+};
+class CLASS(Uniform_Garment_RS_GS_BP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_GS_BP_BB);
+    picture = QPATHTOF(UI\uniform_garment_rs_gs_bp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Garment_RS_GS_BP_BB);
+    };
+};
+class CLASS(Uniform_Garment_RS_GS_EP_TB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_GS_EP_TB);
+    picture = QPATHTOF(UI\uniform_garment_rs_gs_ep_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Garment_RS_GS_EP_TB);
+    };
+};
+class CLASS(Uniform_Garment_RS_ES_EP_TB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_ES_EP_TB);
+    picture = QPATHTOF(UI\uniform_garment_rs_es_ep_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Garment_RS_ES_EP_TB);
+    };
+};
+class CLASS(Uniform_Garment_RS_ES_BP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_ES_BP_BB);
+    picture = QPATHTOF(UI\uniform_garment_rs_es_bp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Garment_RS_ES_BP_BB);
+    };
+};
+class CLASS(Uniform_Garment_RS_ES_GP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas";
+    displayName = CSTRING(Uniform_Garment_RS_ES_GP_BB);
+    picture = QPATHTOF(UI\uniform_garment_rs_es_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Garment_RS_ES_GP_BB);
+    };
+};
+class CLASS(Uniform_Garment_RS_TS_TP_TB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_Garment_RS_TS_TP_TB);
+    picture = QPATHTOF(UI\uniform_garment_rs_ts_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Garment_RS_TS_TP_TB);
+    };
+};
+class CLASS(Uniform_Garment_RS_GS_TP_TB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_Garment_RS_GS_TP_TB);
+    picture = QPATHTOF(UI\uniform_garment_rs_gs_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Garment_RS_GS_TP_TB);
+    };
+};
+class CLASS(Uniform_Garment_RS_TS_GP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_Garment_RS_TS_GP_BB);
+    picture = QPATHTOF(UI\uniform_garment_rs_ts_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Garment_RS_TS_GP_BB);
+    };
+};
+class CLASS(Uniform_Garment_RS_OS_TP_TB): CLASS(Uniform_Base) { // O = Brown
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_Garment_RS_OS_TP_TB);
+    picture = QPATHTOF(UI\uniform_garment_rs_os_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Garment_RS_OS_TP_TB);
+    };
+};
+class CLASS(Uniform_Garment_RS_TS_BP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_Garment_RS_TS_BP_BB);
+    picture = QPATHTOF(UI\uniform_garment_rs_ts_bp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Garment_RS_TS_BP_BB);
+    };
+};
