+++ addons/units/CfgWeapons/plaid_jeans_uniforms.hpp
@@ -1,0 +1,72 @@
+// Naming pattern: TypeSleevesOrPants_ColorShirt_ColorPants_ColorBoots
+// Plaid Shirts with Jeans
+class CLASS(Uniform_Plaid_JP_GY_BP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_JP_GY_BP_BB);
+    picture = QPATHTOF(UI\uniform_plaid_jp_gy_bp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Plaid_JP_GY_BP_BB);
+    };
+};
+class CLASS(Uniform_Plaid_JP_LP_BP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_JP_LP_BP_BB);
+    picture = QPATHTOF(UI\uniform_plaid_jp_lp_bp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Plaid_JP_LP_BP_BB);
+    };
+};
+class CLASS(Uniform_Plaid_BP_LPP_BP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_BP_LPP_BP_BB);
+    picture = QPATHTOF(UI\uniform_plaid_bp_lpp_bp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Plaid_BP_LPP_BP_BB);
+    };
+};
+class CLASS(Uniform_Plaid_BP_LY_BP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_BP_LY_BP_BB);
+    picture = QPATHTOF(UI\uniform_plaid_bb_ly_bp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Plaid_BP_LY_BP_BB);
+    };
+};
+class CLASS(Uniform_Plaid_JP_CB_BP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_JP_CB_BP_BB);
+    picture = QPATHTOF(UI\uniform_plaid_jp_cb_bp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Plaid_JP_CB_BP_BB);
+    };
+};
+class CLASS(Uniform_Plaid_JP_RO_BP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_JP_RO_BP_BB);
+    picture = QPATHTOF(UI\uniform_plaid_jp_ro_bp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Plaid_JP_RO_BP_BB);
+    };
+};
+class CLASS(Uniform_Plaid_JP_RT_BP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_JP_RT_BP_BB);
+    picture = QPATHTOF(UI\uniform_plaid_jp_rt_bp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Plaid_JP_RT_BP_BB);
+    };
+};
