+++ addons/units/CfgWeapons/plaid_shorts_uniforms.hpp
@@ -1,0 +1,72 @@
+// Naming pattern: TypeSleevesOrPants_ColorShirt_ColorPants_ColorBoots
+// Plaid Shirts with Shorts
+class CLASS(Uniform_Plaid_SP_GY_BP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_SP_GY_BP_BB);
+    picture = QPATHTOF(UI\uniform_plaid_sp_gy_bp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Plaid_SP_GY_BP_BB);
+    };
+};
+class CLASS(Uniform_Plaid_SP_LP_BP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_SP_LP_BP_BB);
+    picture = QPATHTOF(UI\uniform_plaid_sp_lp_bp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Plaid_SP_LP_BP_BB);
+    };
+};
+class CLASS(Uniform_Plaid_SP_LPP_BP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_SP_LPP_BP_BB);
+    picture = QPATHTOF(UI\uniform_plaid_sp_lpp_bp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Plaid_SP_LPP_BP_BB);
+    };
+};
+class CLASS(Uniform_Plaid_SP_LY_BP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_SP_LY_BP_BB);
+    picture = QPATHTOF(UI\uniform_plaid_sp_ly_bp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Plaid_SP_LY_BP_BB);
+    };
+};
+class CLASS(Uniform_Plaid_SP_CB_BP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_SP_CB_BP_BB);
+    picture = QPATHTOF(UI\uniform_plaid_sp_cb_bp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Plaid_SP_CB_BP_BB);
+    };
+};
+class CLASS(Uniform_Plaid_SP_RO_BP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_SP_RO_BP_BB);
+    picture = QPATHTOF(UI\uniform_plaid_sp_ro_bp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Plaid_SP_RO_BP_BB);
+    };
+};
+class CLASS(Uniform_Plaid_SP_RT_BP_BB): CLASS(Uniform_Base_Floral) {
+    scope = 2;
+    author = "Trenchgun, Tyrone";
+    displayName = CSTRING(Uniform_Plaid_SP_RT_BP_BB);
+    picture = QPATHTOF(UI\uniform_plaid_sp_rt_bp_bb_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Plaid_SP_RT_BP_BB);
+    };
+};
