+++ addons/units/CfgWeapons/polo_tactical_uniforms.hpp
@@ -1,0 +1,222 @@
+// Naming pattern: TypeSleevesOrPants_ColorShirt_ColorPants_ColorBoots
+// Polo - Tactical Pants
+class CLASS(Uniform_Polo_TP_LS_TP_TB): CLASS(Uniform_Base) { // L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_Polo_TP_LS_TP_TB);
+    picture = QPATHTOF(UI\uniform_polo_tp_ls_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_LS_TP_TB);
+    };
+};
+class CLASS(Uniform_Polo_TP_LS_TP_TB_NoLogo): CLASS(Uniform_Base) { // L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Kresky";
+    displayName = CSTRING(Uniform_Polo_TP_LS_TP_TB_NoLogo);
+    picture = QPATHTOF(UI\uniform_polo_tp_ls_tp_nologo_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_LS_TP_TB_NoLogo);
+    };
+};
+class CLASS(Uniform_Polo_TP_TS_GP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_Polo_TP_TS_GP_BB);
+    picture = QPATHTOF(UI\uniform_polo_tp_ts_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_TS_GP_BB);
+    };
+};
+class CLASS(Uniform_Polo_TP_TS_GP_BB_NoLogo): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas, Kresky";
+    displayName = CSTRING(Uniform_Polo_TP_TS_GP_BB_NoLogo);
+    picture = QPATHTOF(UI\uniform_polo_tp_ts_gp_nologo_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_TS_GP_BB_NoLogo);
+    };
+};
+class CLASS(Uniform_Polo_TP_BS_TP_TB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_Polo_TP_BS_TP_TB);
+    picture = QPATHTOF(UI\uniform_polo_tp_bs_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_BS_TP_TB);
+    };
+};
+class CLASS(Uniform_Polo_TP_BS_TP_TB_NoLogo): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas, Kresky";
+    displayName = CSTRING(Uniform_Polo_TP_BS_TP_TB_NoLogo);
+    picture = QPATHTOF(UI\uniform_polo_tp_bs_tp_nologo_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_BS_TP_TB_NoLogo);
+    };
+};
+class CLASS(Uniform_Polo_TP_BS_LP_BB): CLASS(Uniform_Base) { // L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_Polo_TP_BS_LP_BB);
+    picture = QPATHTOF(UI\uniform_polo_tp_bs_lp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_BS_LP_BB);
+    };
+};
+class CLASS(Uniform_Polo_TP_BS_LP_BB_NoLogo): CLASS(Uniform_Base) { // L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Kresky";
+    displayName = CSTRING(Uniform_Polo_TP_BS_LP_BB_NoLogo);
+    picture = QPATHTOF(UI\uniform_polo_tp_bs_lp_nologo_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_BS_LP_BB_NoLogo);
+    };
+};
+class CLASS(Uniform_Polo_TP_LS_GP_BB): CLASS(Uniform_Base) { // L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_Polo_TP_LS_GP_BB);
+    picture = QPATHTOF(UI\uniform_polo_tp_ls_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_LS_GP_BB);
+    };
+};
+class CLASS(Uniform_Polo_TP_LS_GP_BB_NoLogo): CLASS(Uniform_Base) { // L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Kresky";
+    displayName = CSTRING(Uniform_Polo_TP_LS_GP_BB_NoLogo);
+    picture = QPATHTOF(UI\uniform_polo_tp_ls_gp_nologo_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_LS_GP_BB_NoLogo);
+    };
+};
+class CLASS(Uniform_Polo_TP_OS_TP_BB): CLASS(Uniform_Base) { // O = Brown
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_Polo_TP_OS_TP_BB);
+    picture = QPATHTOF(UI\uniform_polo_tp_os_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_OS_TP_BB);
+    };
+};
+class CLASS(Uniform_Polo_TP_OS_TP_BB_NoLogo): CLASS(Uniform_Base) { // O = Brown
+    scope = 2;
+    author = "Pomigit, Jonpas, Kresky";
+    displayName = CSTRING(Uniform_Polo_TP_OS_TP_BB_NoLogo);
+    picture = QPATHTOF(UI\uniform_polo_tp_os_tp_nologo_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_OS_TP_BB_NoLogo);
+    };
+};
+class CLASS(Uniform_Polo_TP_OS_LP_BB): CLASS(Uniform_Base) { // O = Brown, L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_Polo_TP_OS_LP_BB);
+    picture = QPATHTOF(UI\uniform_polo_tp_os_lp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_OS_LP_BB);
+    };
+};
+class CLASS(Uniform_Polo_TP_OS_LP_BB_NoLogo): CLASS(Uniform_Base) { // O = Brown, L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Kresky";
+    displayName = CSTRING(Uniform_Polo_TP_OS_LP_BB_NoLogo);
+    picture = QPATHTOF(UI\uniform_polo_tp_os_lp_nologo_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_OS_LP_BB_NoLogo);
+    };
+};
+class CLASS(Uniform_Polo_TP_GS_TP_TB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_Polo_TP_GS_TP_TB);
+    picture = QPATHTOF(UI\uniform_polo_tp_gs_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_GS_TP_TB);
+    };
+};
+class CLASS(Uniform_Polo_TP_GS_TP_TB_NoLogo): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas, Kresky";
+    displayName = CSTRING(Uniform_Polo_TP_GS_TP_TB_NoLogo);
+    picture = QPATHTOF(UI\uniform_polo_tp_gs_tp_nologo_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_GS_TP_TB_NoLogo);
+    };
+};
+class CLASS(Uniform_Polo_TP_WS_TP_TB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_Polo_TP_WS_TP_TB);
+    picture = QPATHTOF(UI\uniform_polo_tp_ws_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_WS_TP_TB);
+    };
+};
+class CLASS(Uniform_Polo_TP_WS_TP_TB_NoLogo): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas, Kresky";
+    displayName = CSTRING(Uniform_Polo_TP_WS_TP_TB_NoLogo);
+    picture = QPATHTOF(UI\uniform_polo_tp_ws_tp_nologo_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_WS_TP_TB_NoLogo);
+    };
+};
+class CLASS(Uniform_Polo_TP_WS_LP_BB): CLASS(Uniform_Base) { // L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_Polo_TP_WS_LP_BB);
+    picture = QPATHTOF(UI\uniform_polo_tp_ws_lp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_WS_LP_BB);
+    };
+};
+class CLASS(Uniform_Polo_TP_WS_LP_BB_NoLogo): CLASS(Uniform_Base) { // L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Kresky";
+    displayName = CSTRING(Uniform_Polo_TP_WS_LP_BB_NoLogo);
+    picture = QPATHTOF(UI\uniform_polo_tp_ws_lp_nologo_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_WS_LP_BB_NoLogo);
+    };
+};
+class CLASS(Uniform_Polo_TP_WS_GP_BB): CLASS(Uniform_Base) { // L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_Polo_TP_WS_GP_BB);
+    picture = QPATHTOF(UI\uniform_polo_tp_ws_gp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_WS_GP_BB);
+    };
+};
+class CLASS(Uniform_Polo_TP_WS_GP_BB_NoLogo): CLASS(Uniform_Base) { // L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Kresky";
+    displayName = CSTRING(Uniform_Polo_TP_WS_GP_BB_NoLogo);
+    picture = QPATHTOF(UI\uniform_polo_tp_ws_gp_nologo_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_Polo_TP_WS_GP_BB_NoLogo);
+    };
+};
