+++ addons/units/CfgWeapons/tshirt_jeans_uniforms.hpp
@@ -1,0 +1,62 @@
+// Naming pattern: TypeSleevesOrPants_ColorShirt_ColorPants_ColorBoots
+// TShirt - Jean Pants
+class CLASS(Uniform_TShirt_JP_GS_LP_BB): CLASS(Uniform_Base) { // L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_TShirt_JP_GS_LP_BB);
+    picture = QPATHTOF(UI\uniform_tshirt_jp_gs_lp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_TShirt_JP_GS_LP_BB);
+    };
+};
+class CLASS(Uniform_TShirt_JP_GS_TP_BB): CLASS(Uniform_Base) { // L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_TShirt_JP_GS_TP_BB);
+    picture = QPATHTOF(UI\uniform_tshirt_jp_gs_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_TShirt_JP_GS_TP_BB);
+    };
+};
+class CLASS(Uniform_TShirt_JP_BS_LP_BB): CLASS(Uniform_Base) { // L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_TShirt_JP_BS_LP_BB);
+    picture = QPATHTOF(UI\uniform_tshirt_jp_bs_lp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_TShirt_JP_BS_LP_BB);
+    };
+};
+class CLASS(Uniform_TShirt_JP_BS_TP_BB): CLASS(Uniform_Base) {
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_TShirt_JP_BS_TP_BB);
+    picture = QPATHTOF(UI\uniform_tshirt_jp_bs_tp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_TShirt_JP_BS_TP_BB);
+    };
+};
+class CLASS(Uniform_TShirt_JP_LS_LP_BB): CLASS(Uniform_Base) { // L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_TShirt_JP_LS_LP_BB);
+    picture = QPATHTOF(UI\uniform_tshirt_jp_ls_lp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_TShirt_JP_LS_LP_BB);
+    };
+};
+class CLASS(Uniform_TShirt_JP_WS_LP_BB): CLASS(Uniform_Base) { // L = Blue
+    scope = 2;
+    author = "Pomigit, Jonpas, Rory";
+    displayName = CSTRING(Uniform_TShirt_JP_WS_LP_BB);
+    picture = QPATHTOF(UI\uniform_tshirt_jp_ws_lp_ca.paa);
+
+    class ItemInfo: ItemInfo {
+        uniformClass = QCLASS(Unit_TShirt_JP_WS_LP_BB);
+    };
+};
