+++ addons/units/stringtable.xml
@@ -19,18 +19,6 @@
         <Key ID="STR_TACS_Units_Uniform_Combat_LS_ATACS_LEX">
             <English>Combat Uniform Long (A-TACS LE-X)</English>
         </Key>
-        <Key ID="STR_TACS_Units_Uniform_Combat_RS_ATACS_ATX">
-            <English>Combat Uniform Rolled (A-TACS A-TX)</English>
-        </Key>
-        <Key ID="STR_TACS_Units_Uniform_Combat_RS_ATACS_AUX">
-            <English>Combat Uniform Rolled (A-TACS AU-X)</English>
-        </Key>
-        <Key ID="STR_TACS_Units_Uniform_Combat_RS_ATACS_IX">
-            <English>Combat Uniform Rolled (A-TACS iX)</English>
-        </Key>
-        <Key ID="STR_TACS_Units_Uniform_Combat_RS_ATACS_LEX">
-            <English>Combat Uniform Rolled (A-TACS LE-X)</English>
-        </Key>
         <Key ID="STR_TACS_Units_Uniform_Combat_LS_BS_BP_BB">
             <English>Combat Uniform Long (Black, Black, Black)</English>
             <German>Kampfuniform, lang (Schwarz, Schwarz, Schwarz)</German>
@@ -158,6 +146,18 @@
             <German>Kampfuniform, lang (Weiß, Weiß, Weiß)</German>
             <Japanese>戦闘服 長袖 (白、白、白)</Japanese>
             <French>Uniforme de combat long (Blanc, Blanc, Blanc)</French>
+        </Key>
+        <Key ID="STR_TACS_Units_Uniform_Combat_RS_ATACS_ATX">
+            <English>Combat Uniform Rolled (A-TACS A-TX)</English>
+        </Key>
+        <Key ID="STR_TACS_Units_Uniform_Combat_RS_ATACS_AUX">
+            <English>Combat Uniform Rolled (A-TACS AU-X)</English>
+        </Key>
+        <Key ID="STR_TACS_Units_Uniform_Combat_RS_ATACS_IX">
+            <English>Combat Uniform Rolled (A-TACS iX)</English>
+        </Key>
+        <Key ID="STR_TACS_Units_Uniform_Combat_RS_ATACS_LEX">
+            <English>Combat Uniform Rolled (A-TACS LE-X)</English>
         </Key>
         <Key ID="STR_TACS_Units_Uniform_Combat_RS_BS_BP_BB">
             <English>Combat Uniform Rolled (Black, Black, Black)</English>
